Maintain Python comfort models - rename module to comfort_models - style (whitespace, long lines) - remove class container
